Extract Git repo file content replacement logic into a facet method
Replacing the content of a file in Git is non-trivial.  Extract the
logic used to update the index repo's config.json to a reusable method.
